refactor(server): extract model profile helpers from heartbeat.ts
Moves the model-profile cluster (resolution, wake-context
normalization, adapter-config merge, run metadata) into
heartbeat-model-profile.ts, following the existing heartbeat-<domain>
sibling pattern. heartbeat.ts re-exports the public symbols so
consumers and tests are unaffected. No behavior change.
